# internal/api/docs/swagger.yaml
definitions:
...
  api.GetProjectOutput:
    properties:
      project:
        $ref: '#/definitions/db.Project'
    type: object
  api.GetProjectsOutput:
    properties:
      projects:
        items:
          $ref: '#/definitions/db.Project'
        type: array
    type: object
...
  api.UpsertProjectInput:
    properties:
      description:
        type: string
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
      type:
        type: string
      year:
        type: string
    required:
    - description
    - end_date
    - name
    - start_date
    - type
    - year
    type: object
...
  db.Project:
    properties:
      created:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      name:
        type: string
      start_date:
        type: string
      type:
        type: string
      updated:
        type: string
      userid:
        type: string
      year:
        type: string
    type: object
...
info:
  contact: {}
  title: 4H Record Books API
paths:
...
  /project:
    get:
      consumes:
      - application/json
      description: Gets all of a user's saved projects regardless of year
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetProjectsOutput'
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Get all of a user's projects
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Adds a project to a user's personal records
      parameters:
      - description: Project information
        in: body
        name: UpsertProjectInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertProjectInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Add a project
      tags:
      - Project
  /project/{projectId}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's project given the project ID
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Removes a project
      tags:
      - Project
    get:
      consumes:
      - application/json
      description: Get a user's project by ID
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetProjectOutput'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get a project
      tags:
      - Project
    put:
      consumes:
      - application/json
      description: Updates a user's project information
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Project information
        in: body
        name: UpsertProjectInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertProjectInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Update a project
      tags:
      - Project
  /projects:
    get:
      consumes:
      - application/json
      description: Gets all of a user's projects that take place in the last 12 months
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetProjectsOutput'
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Gets projects of the current year
      tags:
      - Project
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"